add operations.py; self check passes
