+++ pyanalyze/test_value.py
@@ -31,10 +31,8 @@ from .value import (
     MultiValuedValue,
     SubclassValue,
     CanAssignContext,
-    SequenceIncompleteValue,
     TypeVarMap,
     UNRESOLVED_VALUE,
-    concrete_values_from_iterable,
 )
 
 
@@ -175,14 +173,21 @@ def test_protocol() -> None:
 
 
 def test_protocol_value() -> None:
-    pv = ProtocolValue("Proto", members={"numerator": (TypedValue(int), {})})
+    pv = ProtocolValue(
+        "pyanalyze.test_value", "Proto", members={"numerator": (TypedValue(int), {})}
+    )
     assert_can_assign(pv, pv)
     assert_can_assign(pv, TypedValue(int))
     assert_can_assign(pv, KnownValue(1))
     assert_cannot_assign(pv, TypedValue(str))
     assert_cannot_assign(pv, KnownValue("x"))
     assert_cannot_assign(
-        pv, ProtocolValue("Proto", members={"numerator": (TypedValue(str), {})})
+        pv,
+        ProtocolValue(
+            "pyanalyze.test_value",
+            "Proto2",
+            members={"numerator": (TypedValue(str), {})},
+        ),
     )
 
 
@@ -425,61 +430,6 @@ def test_annotated_value() -> None:
 def test_io() -> None:
     assert_can_assign(
         GenericValue(typing.IO, [UNRESOLVED_VALUE]), TypedValue(io.BytesIO)
-    )
-
-
-def test_concrete_values_from_iterable() -> None:
-    assert_is(None, concrete_values_from_iterable(KnownValue(1), CTX))
-    assert_eq((), concrete_values_from_iterable(KnownValue(()), CTX))
-    assert_eq(
-        (KnownValue(1), KnownValue(2)),
-        concrete_values_from_iterable(KnownValue((1, 2)), CTX),
-    )
-    assert_eq(
-        MultiValuedValue((KnownValue(1), KnownValue(2))),
-        concrete_values_from_iterable(
-            SequenceIncompleteValue(list, [KnownValue(1), KnownValue(2)]), CTX
-        ),
-    )
-    assert_eq(
-        TypedValue(int),
-        concrete_values_from_iterable(GenericValue(list, [TypedValue(int)]), CTX),
-    )
-    assert_eq(
-        MultiValuedValue([KnownValue(1), KnownValue(3), KnownValue(2), KnownValue(4)]),
-        concrete_values_from_iterable(
-            MultiValuedValue(
-                [
-                    SequenceIncompleteValue(list, [KnownValue(1), KnownValue(2)]),
-                    KnownValue((3, 4)),
-                ]
-            ),
-            CTX,
-        ),
-    )
-    assert_eq(
-        MultiValuedValue([KnownValue(1), KnownValue(2), TypedValue(int)]),
-        concrete_values_from_iterable(
-            MultiValuedValue(
-                [
-                    SequenceIncompleteValue(list, [KnownValue(1), KnownValue(2)]),
-                    GenericValue(list, [TypedValue(int)]),
-                ]
-            ),
-            CTX,
-        ),
-    )
-    assert_eq(
-        MultiValuedValue([KnownValue(1), KnownValue(2), KnownValue(3)]),
-        concrete_values_from_iterable(
-            MultiValuedValue(
-                [
-                    SequenceIncompleteValue(list, [KnownValue(1), KnownValue(2)]),
-                    KnownValue((3,)),
-                ]
-            ),
-            CTX,
-        ),
     )
